requirements.txt
# Test stack
pytest==8.3.3
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-excel==1.7.0
pytest-playwright==0.5.0

# Playwright core (khóa theo version browsers đã build)
playwright==1.46.0
pytest-playwright==0.5.0

# Tiện ích
python-dotenv==1.0.1
openpyxl>=3.1.2
pandas>=2.0
